+++ src/ButtonDropdown.php
@@ -10,6 +10,7 @@ declare(strict_types=1);
 namespace yii\bootstrap5;
 
 use Throwable;
+use Yii;
 use yii\helpers\ArrayHelper;
 use yii\helpers\Url;
 
@@ -54,9 +55,9 @@ class ButtonDropdown extends Widget
     const DIRECTION_UP = 'up';
 
     /**
-     * @var string the button label
+     * @var string|null the button label
      */
-    public $label = 'Button';
+    public $label = null;
     /**
      * @var array the HTML attributes for the container tag. The following special options are recognized:
      *
@@ -113,6 +114,9 @@ class ButtonDropdown extends Widget
         if (!isset($this->buttonOptions['id'])) {
             $this->buttonOptions['id'] = $this->options['id'] . '-button';
         }
+        if ($this->label === null) {
+            $this->label = Yii::t('yii/bootstrap5', 'Button');
+        }
     }
 
     /**
@@ -159,7 +163,7 @@ class ButtonDropdown extends Widget
             Html::addCssClass($this->buttonOptions, ['toggle' => 'dropdown-toggle dropdown-toggle-split']);
             unset($buttonOptions['id']);
             $splitButton = Button::widget([
-                'label' => '<span class="visually-hidden">Toggle Dropdown</span>',
+                'label' => '<span class="visually-hidden">' . Yii::t('yii/bootstrap5', 'Toggle Dropdown') . '</span>',
                 'encodeLabel' => false,
                 'options' => $this->buttonOptions,
                 'view' => $this->getView(),