+++ src/Modal.php
@@ -9,6 +9,7 @@ declare(strict_types=1);
 
 namespace yii\bootstrap5;
 
+use Yii;
 use yii\base\InvalidConfigException;
 use yii\helpers\ArrayHelper;
 
@@ -290,7 +291,7 @@ class Modal extends Widget
             $this->closeButton = array_merge([
                 'class' => ['widget' => 'btn-close'],
                 'data' => ['bs-dismiss' => 'modal'],
-                'aria' => ['label' => 'Close']
+                'aria' => ['label' => Yii::t('yii/bootstrap5', 'Close')]
             ], $this->closeButton);
         }